Extract IFittingTest from IMinimizerTest
Redmine: #1859

and add automatic memory management for the vector of FitParameterPlan
