+++ Tests/Functional/Fit/IMinimizerTest.h
@@ -16,40 +16,21 @@
 #ifndef IMINIMIZERTEST_H
 #define IMINIMIZERTEST_H
 
-#include "IFunctionalTest.h"
-#include "OutputData.h"
-#include "FitParameterPlan.h"
-#include <memory>
-
-class FitSuite;
-class MultiLayer;
-class Simulation;
+#include "IFittingTest.h"
 
 //! @class IMinimizerTest
 //! @ingroup standard_samples
 //! @brief Basic class for all minimizer functional tests
 
-class IMinimizerTest : public IFunctionalTest
+class IMinimizerTest : public IFittingTest
 {
 public:
     IMinimizerTest(const std::string &minimizer_name,
                    const std::string &minimizer_algorithm = std::string());
     virtual ~IMinimizerTest();
 
-    bool runTest() final;
-
 protected:
-    virtual void initParameterPlan();
-    virtual std::unique_ptr<FitSuite> createFitSuite();
-    virtual std::unique_ptr<MultiLayer> createSample();
-    virtual std::unique_ptr<Simulation> createSimulation();
-    virtual std::unique_ptr<OutputData<double>> createOutputData(const Simulation* simulation);
-
-    std::vector<FitParameterPlan*> m_parplans;
-    std::string m_minimizer_name;
-    std::string m_minimizer_algorithm;
-    std::string m_simulation_name;
-    std::string m_sample_builder_name;
+    virtual void initParameterPlan() override;
 };
 
 #endif // IMINIMIZERTEST_H
